initial refactor, remove preventPlayerIdentifying
